Fix | #4 | @lcomment | 통합 테스트 추상 클래스 수정
