--- cakk-api/src/test/java/com/cakk/api/common/base/ControllerTest.java
@@ -1,61 +1,0 @@
-package com.cakk.api.common.base;
-
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.filter.CharacterEncodingFilter;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.navercorp.fixturemonkey.FixtureMonkey;
-import com.navercorp.fixturemonkey.api.introspector.BuilderArbitraryIntrospector;
-import com.navercorp.fixturemonkey.api.introspector.ConstructorPropertiesArbitraryIntrospector;
-import com.navercorp.fixturemonkey.api.introspector.FieldReflectionArbitraryIntrospector;
-
-@AutoConfigureMockMvc
-@ActiveProfiles("test")
-@SpringBootTest(properties = "spring.profiles.active: test")
-public abstract class ControllerTest {
-
-	@Autowired
-	protected MockMvc mockMvc;
-
-	@Autowired
-	protected ObjectMapper objectMapper;
-
-	@BeforeEach
-	void setup(WebApplicationContext webApplicationContext) {
-		mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-			.addFilters(new CharacterEncodingFilter("UTF-8", true))
-			.alwaysDo(print())
-			.build();
-	}
-
-	protected final FixtureMonkey getConstructorMonkey() {
-		return FixtureMonkey.builder()
-			.objectIntrospector(ConstructorPropertiesArbitraryIntrospector.INSTANCE)
-			.build();
-	}
-
-	protected final FixtureMonkey getReflectionMonkey() {
-		return FixtureMonkey.builder()
-			.objectIntrospector(FieldReflectionArbitraryIntrospector.INSTANCE)
-			.build();
-	}
-
-	protected final FixtureMonkey getBuilderMonkey() {
-		return FixtureMonkey.builder()
-			.objectIntrospector(BuilderArbitraryIntrospector.INSTANCE)
-			.build();
-	}
-
-	protected String toJson(Object request) throws JsonProcessingException {
-		return objectMapper.writeValueAsString(request);
-	}
-}
