+++ cakk-api/src/test/java/com/cakk/api/common/base/IntegrationTest.java
@@ -1,0 +1,40 @@
+package com.cakk.api.common.base;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.test.context.ActiveProfiles;
+
+import com.navercorp.fixturemonkey.FixtureMonkey;
+import com.navercorp.fixturemonkey.api.introspector.BuilderArbitraryIntrospector;
+import com.navercorp.fixturemonkey.api.introspector.ConstructorPropertiesArbitraryIntrospector;
+import com.navercorp.fixturemonkey.api.introspector.FieldReflectionArbitraryIntrospector;
+
+@ActiveProfiles("test")
+@SpringBootTest(
+	properties = "spring.profiles.active: test",
+	webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
+)
+public abstract class IntegrationTest {
+
+	@Autowired
+	protected TestRestTemplate restTemplate;
+
+	protected final FixtureMonkey getConstructorMonkey() {
+		return FixtureMonkey.builder()
+			.objectIntrospector(ConstructorPropertiesArbitraryIntrospector.INSTANCE)
+			.build();
+	}
+
+	protected final FixtureMonkey getReflectionMonkey() {
+		return FixtureMonkey.builder()
+			.objectIntrospector(FieldReflectionArbitraryIntrospector.INSTANCE)
+			.build();
+	}
+
+	protected final FixtureMonkey getBuilderMonkey() {
+		return FixtureMonkey.builder()
+			.objectIntrospector(BuilderArbitraryIntrospector.INSTANCE)
+			.build();
+	}
+}
